requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
dj-database-url==2.3.0
    # via -r requirements.in
django==5.2.1
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-cors-headers==4.7.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
drf-yasg==1.21.10
    # via -r requirements.in
et-xmlfile==2.0.0
    # via openpyxl
inflection==0.5.1
    # via drf-yasg
numpy==2.2.6
    # via pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via drf-yasg
pandas==2.2.3
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via
    #   drf-yasg
    #   pandas
pyyaml==6.0.2
    # via drf-yasg
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.13.2
    # via dj-database-url
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via drf-yasg
